LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-max-width-interpolation-expected.txt


PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (-0.3) should be [initial]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (0) should be [initial]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (0.3) should be [initial]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (-0.3) should be [initial]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (0) should be [initial]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (0.3) should be [initial]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (0) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (0.3) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions: property <max-width> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (0) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (-0.3) should be [initial]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (0) should be [initial]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (0.3) should be [initial]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (0.5) should be [20px]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (0.6) should be [20px]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (1) should be [20px]
PASS CSS Animations: property <max-width> from [initial] to [20px] at (1.5) should be [20px]
PASS Web Animations: property <max-width> from [initial] to [20px] at (-0.3) should be [initial]
PASS Web Animations: property <max-width> from [initial] to [20px] at (0) should be [initial]
PASS Web Animations: property <max-width> from [initial] to [20px] at (0.3) should be [initial]
PASS Web Animations: property <max-width> from [initial] to [20px] at (0.5) should be [20px]
PASS Web Animations: property <max-width> from [initial] to [20px] at (0.6) should be [20px]
PASS Web Animations: property <max-width> from [initial] to [20px] at (1) should be [20px]
PASS Web Animations: property <max-width> from [initial] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (-0.3) should be [unset]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (0) should be [unset]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (0.3) should be [unset]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (-0.3) should be [unset]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (0) should be [unset]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (0.3) should be [unset]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (0) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (0.3) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions: property <max-width> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (0) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (-0.3) should be [unset]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (0) should be [unset]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (0.3) should be [unset]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (0.5) should be [20px]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (0.6) should be [20px]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (1) should be [20px]
PASS CSS Animations: property <max-width> from [unset] to [20px] at (1.5) should be [20px]
PASS Web Animations: property <max-width> from [unset] to [20px] at (-0.3) should be [unset]
PASS Web Animations: property <max-width> from [unset] to [20px] at (0) should be [unset]
PASS Web Animations: property <max-width> from [unset] to [20px] at (0.3) should be [unset]
PASS Web Animations: property <max-width> from [unset] to [20px] at (0.5) should be [20px]
PASS Web Animations: property <max-width> from [unset] to [20px] at (0.6) should be [20px]
PASS Web Animations: property <max-width> from [unset] to [20px] at (1) should be [20px]
PASS Web Animations: property <max-width> from [unset] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (-0.3) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (0) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (0.3) should be [none]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (-0.3) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (0) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (0.3) should be [none]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (0) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (0.3) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions: property <max-width> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (-0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (0) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (0.3) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (1) should be [20px]
PASS CSS Transitions with transition: all: property <max-width> from [none] to [20px] at (1.5) should be [20px]
PASS CSS Animations: property <max-width> from [none] to [20px] at (-0.3) should be [none]
PASS CSS Animations: property <max-width> from [none] to [20px] at (0) should be [none]
PASS CSS Animations: property <max-width> from [none] to [20px] at (0.3) should be [none]
PASS CSS Animations: property <max-width> from [none] to [20px] at (0.5) should be [20px]
PASS CSS Animations: property <max-width> from [none] to [20px] at (0.6) should be [20px]
PASS CSS Animations: property <max-width> from [none] to [20px] at (1) should be [20px]
PASS CSS Animations: property <max-width> from [none] to [20px] at (1.5) should be [20px]
PASS Web Animations: property <max-width> from [none] to [20px] at (-0.3) should be [none]
PASS Web Animations: property <max-width> from [none] to [20px] at (0) should be [none]
PASS Web Animations: property <max-width> from [none] to [20px] at (0.3) should be [none]
PASS Web Animations: property <max-width> from [none] to [20px] at (0.5) should be [20px]
PASS Web Animations: property <max-width> from [none] to [20px] at (0.6) should be [20px]
PASS Web Animations: property <max-width> from [none] to [20px] at (1) should be [20px]
PASS Web Animations: property <max-width> from [none] to [20px] at (1.5) should be [20px]
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0.6) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (1) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (1.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0.6) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (1) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (1.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (-0.3) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (0) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (0.3) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (0.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (0.6) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (1) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [stretch] to [none] at (1.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (-0.3) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0.3) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0.6) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (1) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (1.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (0.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (0.6) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (1) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <max-width> from [stretch] to [none] at (1.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (0) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (0.3) should be [stretch] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (0.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (0.6) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (1) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <max-width> from [stretch] to [none] at (1.5) should be [none] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <max-width> from [fit-content] to [50px] at (-0.3) should be [calc-size(fit-content, -15px + size * 1.3)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions: property <max-width> from [fit-content] to [50px] at (0) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions: property <max-width> from [fit-content] to [50px] at (0.3) should be [calc-size(fit-content, 15px + size * 0.7)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions: property <max-width> from [fit-content] to [50px] at (0.6) should be [calc-size(fit-content, 30px + size * 0.4)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions: property <max-width> from [fit-content] to [50px] at (1) should be [calc-size(fit-content, 50px + size * 0)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions: property <max-width> from [fit-content] to [50px] at (1.5) should be [calc-size(fit-content, 75px + size * -0.5)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions with transition: all: property <max-width> from [fit-content] to [50px] at (-0.3) should be [calc-size(fit-content, -15px + size * 1.3)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions with transition: all: property <max-width> from [fit-content] to [50px] at (0) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions with transition: all: property <max-width> from [fit-content] to [50px] at (0.3) should be [calc-size(fit-content, 15px + size * 0.7)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions with transition: all: property <max-width> from [fit-content] to [50px] at (0.6) should be [calc-size(fit-content, 30px + size * 0.4)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions with transition: all: property <max-width> from [fit-content] to [50px] at (1) should be [calc-size(fit-content, 50px + size * 0)] assert_equals: expected "150px " but got "50px "
FAIL CSS Transitions with transition: all: property <max-width> from [fit-content] to [50px] at (1.5) should be [calc-size(fit-content, 75px + size * -0.5)] assert_equals: expected "150px " but got "50px "
FAIL CSS Animations: property <max-width> from [fit-content] to [50px] at (-0.3) should be [calc-size(fit-content, -15px + size * 1.3)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Animations: property <max-width> from [fit-content] to [50px] at (0) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Animations: property <max-width> from [fit-content] to [50px] at (0.3) should be [calc-size(fit-content, 15px + size * 0.7)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Animations: property <max-width> from [fit-content] to [50px] at (0.6) should be [calc-size(fit-content, 30px + size * 0.4)] assert_equals: expected "150px " but got "50px "
FAIL CSS Animations: property <max-width> from [fit-content] to [50px] at (1) should be [calc-size(fit-content, 50px + size * 0)] assert_equals: expected "150px " but got "50px "
FAIL CSS Animations: property <max-width> from [fit-content] to [50px] at (1.5) should be [calc-size(fit-content, 75px + size * -0.5)] assert_equals: expected "150px " but got "50px "
FAIL Web Animations: property <max-width> from [fit-content] to [50px] at (-0.3) should be [calc-size(fit-content, -15px + size * 1.3)] assert_equals: expected "150px " but got "fit - content "
FAIL Web Animations: property <max-width> from [fit-content] to [50px] at (0) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "fit - content "
FAIL Web Animations: property <max-width> from [fit-content] to [50px] at (0.3) should be [calc-size(fit-content, 15px + size * 0.7)] assert_equals: expected "150px " but got "fit - content "
FAIL Web Animations: property <max-width> from [fit-content] to [50px] at (0.6) should be [calc-size(fit-content, 30px + size * 0.4)] assert_equals: expected "150px " but got "50px "
FAIL Web Animations: property <max-width> from [fit-content] to [50px] at (1) should be [calc-size(fit-content, 50px + size * 0)] assert_equals: expected "150px " but got "50px "
FAIL Web Animations: property <max-width> from [fit-content] to [50px] at (1.5) should be [calc-size(fit-content, 75px + size * -0.5)] assert_equals: expected "150px " but got "50px "
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (-0.3) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (0) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (0.3) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (0.5) should be [min-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (0.6) should be [min-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (1) should be [min-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [max-content] to [min-content] at (1.5) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (-0.3) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (0) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (0.3) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (0.5) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (0.6) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (1) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [max-content] to [min-content] at (1.5) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (-0.3) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (0) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (0.3) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (0.5) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (0.6) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (1) should be [min-content]
PASS CSS Transitions: property <max-width> from [max-content] to [min-content] at (1.5) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (-0.3) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (0) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (0.3) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (0.5) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (0.6) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (1) should be [min-content]
PASS CSS Transitions with transition: all: property <max-width> from [max-content] to [min-content] at (1.5) should be [min-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (-0.3) should be [max-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (0) should be [max-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (0.3) should be [max-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (0.5) should be [min-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (0.6) should be [min-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (1) should be [min-content]
PASS CSS Animations: property <max-width> from [max-content] to [min-content] at (1.5) should be [min-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (-0.3) should be [max-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (0) should be [max-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (0.3) should be [max-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (0.5) should be [min-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (0.6) should be [min-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (1) should be [min-content]
PASS Web Animations: property <max-width> from [max-content] to [min-content] at (1.5) should be [min-content]
PASS CSS Transitions: property <max-width> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -45px + size * 1.3)]
PASS CSS Transitions: property <max-width> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)]
PASS CSS Transitions: property <max-width> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 45px + size * 0.7)]
PASS CSS Transitions: property <max-width> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 90px + size * 0.4)]
PASS CSS Transitions: property <max-width> from [min-content] to neutral at (1) should be [calc-size(min-content, 150px + size * 0)]
PASS CSS Transitions: property <max-width> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 225px + size * -0.5)]
PASS CSS Transitions with transition: all: property <max-width> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -45px + size * 1.3)]
PASS CSS Transitions with transition: all: property <max-width> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)]
PASS CSS Transitions with transition: all: property <max-width> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 45px + size * 0.7)]
PASS CSS Transitions with transition: all: property <max-width> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 90px + size * 0.4)]
PASS CSS Transitions with transition: all: property <max-width> from [min-content] to neutral at (1) should be [calc-size(min-content, 150px + size * 0)]
PASS CSS Transitions with transition: all: property <max-width> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 225px + size * -0.5)]
FAIL CSS Animations: property <max-width> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -45px + size * 1.3)] assert_equals: expected "150px " but got "min - content "
FAIL CSS Animations: property <max-width> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "150px " but got "min - content "
FAIL CSS Animations: property <max-width> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 45px + size * 0.7)] assert_equals: expected "150px " but got "min - content "
PASS CSS Animations: property <max-width> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 90px + size * 0.4)]
PASS CSS Animations: property <max-width> from [min-content] to neutral at (1) should be [calc-size(min-content, 150px + size * 0)]
PASS CSS Animations: property <max-width> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 225px + size * -0.5)]
FAIL Web Animations: property <max-width> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -45px + size * 1.3)] assert_equals: expected "150px " but got "min - content "
FAIL Web Animations: property <max-width> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "150px " but got "min - content "
FAIL Web Animations: property <max-width> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 45px + size * 0.7)] assert_equals: expected "150px " but got "min - content "
PASS Web Animations: property <max-width> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 90px + size * 0.4)]
PASS Web Animations: property <max-width> from [min-content] to neutral at (1) should be [calc-size(min-content, 150px + size * 0)]
PASS Web Animations: property <max-width> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 225px + size * -0.5)]
FAIL CSS Transitions: property <max-width> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 195px + size * -0.3)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions: property <max-width> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 150px + size * 0)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions: property <max-width> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 105px + size * 0.3)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions: property <max-width> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 60px + size * 0.6)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions: property <max-width> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions: property <max-width> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -75px + size * 1.5)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-width> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 195px + size * -0.3)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-width> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 150px + size * 0)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-width> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 105px + size * 0.3)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-width> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 60px + size * 0.6)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-width> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-width> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -75px + size * 1.5)] assert_equals: expected "150px " but got "fit - content "
PASS CSS Animations: property <max-width> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 195px + size * -0.3)]
PASS CSS Animations: property <max-width> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 150px + size * 0)]
PASS CSS Animations: property <max-width> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 105px + size * 0.3)]
FAIL CSS Animations: property <max-width> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 60px + size * 0.6)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Animations: property <max-width> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "fit - content "
FAIL CSS Animations: property <max-width> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -75px + size * 1.5)] assert_equals: expected "150px " but got "fit - content "
PASS Web Animations: property <max-width> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 195px + size * -0.3)]
PASS Web Animations: property <max-width> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 150px + size * 0)]
PASS Web Animations: property <max-width> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 105px + size * 0.3)]
FAIL Web Animations: property <max-width> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 60px + size * 0.6)] assert_equals: expected "150px " but got "fit - content "
FAIL Web Animations: property <max-width> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "150px " but got "fit - content "
FAIL Web Animations: property <max-width> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -75px + size * 1.5)] assert_equals: expected "150px " but got "fit - content "
